requirements.txt
Django
djangorestframework
pillow
psycopg
django-jazzmin
requests
python-dotenv
PyreBase4
firebase-admin
channels==3.0.5